Version 2.0 : PARCIAL1 TERMINADO COMPLETO (REPORTE, CONSULTA, CRUD)

+++ PARCIAL1_POO2_GERSONMENDOZA/src/main/java/com/sv/udb/controlador/AlumnosCtrl.java
@@ -61,7 +61,7 @@ public class AlumnosCtrl {
        
     }
      
-    public Alumnos get(Long empId)
+    public Alumnos get(int empId)
     {
         EntityManagerFactory emf = Persistence.createEntityManagerFactory("POOPU1");
         EntityManager em = emf.createEntityManager();
